apps/shim/pyproject.toml
[project]
name = "rac-shim"
version = "1.0.0"
requires-python = ">=3.12"
dependencies = [
  "starlette>=0.37",
  "uvicorn[standard]>=0.30",
  "gunicorn>=22.0",
  "httpx>=0.27",
  "joserfc>=1.1",
  "asyncpg>=0.29",
  "azure-identity>=1.19",
  "azure-keyvault-keys>=4.8",
  "azure-keyvault-secrets>=4.8",
  "structlog>=24.0",
  "pydantic-settings>=2.4",
  "opentelemetry-sdk>=1.25",
  "opentelemetry-exporter-otlp-proto-grpc>=1.25",
]

[dependency-groups]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "hypothesis>=6.100",
  "testcontainers[postgres]>=4.0",
  "respx>=0.21",
  "freezegun>=1.5",
  "ruff>=0.5",
  "mypy>=1.11",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rac_shim"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "N", "S"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "S105", "S106"]

[tool.mypy]
python_version = "3.12"
strict = true

[[tool.mypy.overrides]]
module = ["asyncpg", "asyncpg.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--strict-markers"
